src/easy_sanic/requirements.txt
sanic==0.8.3
PyJWT==1.7.1
requests==2.21.0
aioredis==1.2.0
opentracing==2.0.0
asyncpg==0.18.3
python-consul==1.1.0
aiohttp==3.5.4